Thao course mgmt

+++ src/components/diag/DiagSignOutForm.js
@@ -6,6 +6,8 @@ import "../../assets/css/card/DiagForm.css";
 import { resetCenterPI } from "../../store/profileCenterSlice";
 import { resetUserPI } from "../../store/profileUserSlice";
 import { setActiveTypeOfTask } from "../../store/listTaskOfCenterAd";
+import { postLogout } from "../../services/authService";
+import { APIStatus } from "../../constants/constants";
 
 const DiagSignOutForm = ({
   isOpen,
@@ -15,17 +17,20 @@ const DiagSignOutForm = ({
 }) => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
-  const handleSignout = () => {
-    try {
-      localStorage.clear();
-      dispatch(resetUserPI());
-      dispatch(resetCenterPI());
-      dispatch(setActiveTypeOfTask("lectures"));
-      setUnreadCount(null);
-      setNotifications([]);
-    } catch (error) {
-    } finally {
-      navigate("/");
+  const handleSignout = async () => {
+    const response = await postLogout();
+    if (response.status === APIStatus.success) {
+      try {
+        localStorage.clear();
+        dispatch(resetUserPI());
+        dispatch(resetCenterPI());
+        dispatch(setActiveTypeOfTask("lectures"));
+        setUnreadCount(null);
+        setNotifications([]);
+      } catch (error) {
+      } finally {
+        navigate("/");
+      }
     }
   };
   if (!isOpen) return null;
